c309: extract RasterTab cluster (RasterTab + TopDocumentsPreview + DocPredictionPanel + PixelDetailCard) to own file (#441 2.1)
Workspace.tsx 4835 LOC -> 3845 LOC (-990 LOC, second-largest single extraction).
Entry chunk 123.60 KB -> 103.57 KB (-20.0 KB, biggest single-tab delta in campaign so far).
27 tabs extracted + lazy.
Removed newly-unused DominantTopicRaster + PickInfo + RoutedPrediction + computeRoutedPrediction imports from Workspace.tsx.
typecheck clean, 44/44 tests pass.
